requirements.txt
numpy
pyqtgraph
PySide6
rplidar
pyserial
